// src/libs/data-access/klass/models/ClassificationVersionResource.ts
export interface ClassificationVersionResource {
    /**
     * 
     * @type {string}
     * @memberof ClassificationVersionResource
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof ClassificationVersionResource
     */
    id?: number;
    /**
     * 
     * @type {Date}
     * @memberof ClassificationVersionResource
     */
    validFrom?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ClassificationVersionResource
     */
    validTo?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ClassificationVersionResource
     */
    lastModified?: Date;
    /**
     * 
     * @type {Array<string>}
     * @memberof ClassificationVersionResource
     */
    published?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof ClassificationVersionResource
     */
    introduction?: string;
    /**
     * 
     * @type {ContactPersonResource}
     * @memberof ClassificationVersionResource
     */
    contactPerson?: ContactPersonResource;
    /**
     * 
     * @type {string}
     * @memberof ClassificationVersionResource
     */
    owningSection?: string;
    /**
     * 
     * @type {string}
     * @memberof ClassificationVersionResource
     */
    legalBase?: string;
    /**
     * 
     * @type {string}
     * @memberof ClassificationVersionResource
     */
    publications?: string;
    /**
     * 
     * @type {string}
     * @memberof ClassificationVersionResource
     */
    derivedFrom?: string;
    /**
     * 
     * @type {Array<CorrespondenceTableSummaryResource>}
     * @memberof ClassificationVersionResource
     */
    correspondenceTables?: Array<CorrespondenceTableSummaryResource>;
    /**
     * 
     * @type {Array<ClassificationVariantSummaryResource>}
     * @memberof ClassificationVersionResource
     */
    classificationVariants?: Array<ClassificationVariantSummaryResource>;
    /**
     * 
     * @type {Array<ChangelogResource>}
     * @memberof ClassificationVersionResource
     */
    changelogs?: Array<ChangelogResource>;
    /**
     * 
     * @type {Array<LevelResource>}
     * @memberof ClassificationVersionResource
     */
    levels?: Array<LevelResource>;
    /**
     * 
     * @type {Array<ClassificationItemResource>}
     * @memberof ClassificationVersionResource
     */
    classificationItems?: Array<ClassificationItemResource>;
    /**
     * 
     * @type {{ [key: string]: Link; }}
     * @memberof ClassificationVersionResource
     */
    links?: { [key: string]: Link; };
    /**
     * 
     * @type {ClassificationVariantSummaryResourceEmbedded}
     * @memberof ClassificationVersionResource
     */
    embedded?: ClassificationVariantSummaryResourceEmbedded;
    /**
     * 
     * @type {PageMetadata}
     * @memberof ClassificationVersionResource
     */
    page?: PageMetadata;
}